always use world coords instead of chunk coords + cleanup

+++ src/main/java/domrbeeson/gamma/event/events/player/PlayerBlockPlaceEvent.java
@@ -8,7 +8,7 @@ public class PlayerBlockPlaceEvent extends BlockChangeEvent {
 
     private final Player player;
 
-    public PlayerBlockPlaceEvent(Player player, Chunk chunk, int x, byte y, int z, byte currentId, byte currentMetadata, byte newId, byte newMetadata, int clickedX, byte clickedY, int clickedZ, boolean tick) {
+    public PlayerBlockPlaceEvent(Player player, Chunk chunk, int x, int y, int z, byte currentId, byte currentMetadata, byte newId, byte newMetadata, int clickedX, byte clickedY, int clickedZ, boolean tick) {
         super(chunk, x, y, z, currentId, currentMetadata, newId, newMetadata, clickedX, clickedY, clickedZ, tick);
         this.player = player;
     }
